[Infra] Backend 배포 1.1

See merge request s11-fintech-finance-sub1/S11P21B302!81

+++ wannaeat-be/src/main/java/com/waterdragon/wannaeat/domain/reservation/domain/ReservationTable.java
@@ -33,7 +33,6 @@ public class ReservationTable {
 	private Reservation reservation;
 
 	@Column(name = "table_id", nullable = false)
-	private Long tableId;
+	private int tableId;
 
-	// 생성자, getter, setter
 }
